Finish grid layout
The html file is almost complete. It may be removed some button elements
to place more images as icons (e.g. a search or notifition icon). It
also contains dummy text that probably be removed later.

The css uses grid to layout everything on the page. It needs style for
certain divs (the cards). The colors will be changed also. And new fonts
to give it another style.